Change id to uint8_t pointer in spi_flash driver. Set internal Pull up resistor on MISO pin for stm32wb55xx MCU

+++ drivers/memory/spi_flash.h
@@ -17,7 +17,7 @@ typedef enum spi_flash_ioctl_command {
 } spi_flash_ioctl_command_t;
 
 typedef struct spi_flash_ioctl_read_id_arg {
-    uint32_t *id;
+    uint8_t *id;
 } spi_flash_ioctl_read_id_arg_t;
 
 typedef struct spi_flash_ioctl_read_arg {
